requirements.txt
Flask==2.3.3
requests==2.31.0
plotly==5.15.0
pandas==2.0.3
python-telegram-bot==20.7
gunicorn==21.2.0  # Para servir la aplicación Flask en producción


